+++ tests/TestArtworkNormalizer.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "BaseTest.h"
+
+class TestArtworkNormalizer : public BaseTest {
+    Q_OBJECT
+
+private slots:
+    void testCropsSymmetricBordersAroundSquareArtwork();
+    void testPreservesGenuineLandscapeArtwork();
+    void testNormalizesPngInPlace();
+};